// src/app/models/param-role.ts
export class ParamRole {
    age: number;
    growth: number;
    clothsize: number;
    chestsize: number;
    waistsize: number;
    thighsize: number;
    footsize: number;
    necksize: number;
    headsize: number;
    tatu: boolean;
    piercing: boolean;
    appearancetype : {'name':string};
    haircolor: {'name':string};
    city: {'name':string};
    eyecolor: {'name':string};
    gender: {'name':string};
    country: {'name':string};
    employmenttype: {'name':string};
}